rping: revert commit 93635fa33b41d356fa096242fec4ce788194b42f
From Steve Wise:
The changes I added to rping to fix problems I was seeing when running
rping over the upstream iw_cxgb3 have been included in OFED-1.5.3.
That change, however breaks rping over iw_cxgb3 in 1.5.3.
It causes a hang at the end of the rping run.  The problem is not
with rping, but with the down-level iw_cxgb3 code.  The upstream
change to fix iw_cxgb3, however, isn't trivial and it bumps the
cxgb3 uverbs ABI.

To fix this for OFED 1.5.3, we revert the corresponding commit
and release an updated package.

Signed-off-by: Sean Hefty <sean.hefty@intel.com>